requirements.txt
flask==1.0.2
gunicorn==19.9.0
gevent==1.3.7
mysql-connector-python==8.0.12
requests==2.21.0
tqdm==4.23.4
psutil==5.4.8
